fix max items logic
previously, I just lazily did an increment modulo max items. this is
obviously terrible because it ends up such that anything you copy past
max items keeps overwriting entry number 1... now I have fixed it so
that for example with max items = 20, when you copy 21st item, it keeps
it at number 21 itself, and then deletes item number 1. so the list
becomes 2..21, and then it becomes 3...22, and so on.