minor improvement to target parsing logic
some programs expose a target MIME like text/plain;charset=utf-8, and
this is the only text format target they expose. in this case, the
target gets completely butchered by existing parser as it makes it
text/plaincharsetutf-8, which is wrong.

this addition canonicalizes it so that anything after the `;` in the
mime gets chucked, and all such targets resolve to text/plain.