add keyboard actions
Previously, the view was structured like this:

| Window -- handles esc alone and otherwise passes keyboard events through
| ...
|     GtkListBox
|         GtkListBoxRow
|             GtkEventBox -- handles all the mouse interactions
|                 [ actual items ]
|        ...

This is because `GtkListBoxRow` doesn't fire button-press or
button-release buttons.. for some reason. So there was an eventbox
inside that handled all the interactions.

Since we need to add keyboard actions, I thought just adding it to the
event box would work fine. However, this leads to an issue with
tab-navigation. Due to the `GTKListBoxRow` and the `GtkEventBox` both being
focusable, it takes 2 pressed of Tab to actually move from one row to
the next. If you try to mark the `GtkListBoxRow` as "not for tab
navigation", then you cannot navigate to its child `GtkEventBox` either.
There is a way to manually exclude it using `set_focus_chain` where you
iterate down the children and make the focus chain array yourself and
pass it to GTK. While this would have been painful, it is deprecated
anyway, and the alternative proposed in the documentation is
undocumented. Sigh.

However, if we mark the `GtkEventBox` as "not for tab navigation", while
allowing the `GtkListBoxRow` for tab navigation, then it works as
intended.

So, we now have the `GtkEventBox` listening to clicks... and the
`GtkListBoxRow` handling keyboard events.

As long as it works.